updated sort block
