+++ apps/web/app/components/blocks/SortFilter/__tests__/SortFilter.spec.ts
@@ -2,6 +2,12 @@ import { mount } from '@vue/test-utils';
 import SortFilter from '../../../../components/blocks/SortFilter/SortFilter.vue';
 import type { SortFilterProps } from '../types';
 
+vi.mock('vue-router', () => ({
+  useRoute: () => ({
+    query: {},
+  }),
+}));
+
 const mockProps: SortFilterProps = {
   name: 'SortFilter',
   type: 'content',
